Setup launch for CWL to allow imports
