+++ dockstore-client/src/main/java/io/dockstore/client/cli/nested/AbstractEntryClient.java
@@ -17,25 +17,41 @@
 package io.dockstore.client.cli.nested;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
 import org.apache.commons.csv.CSVFormat;
+import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVPrinter;
+import org.apache.commons.csv.CSVRecord;
+import org.apache.commons.csv.QuoteMode;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 
 import com.google.common.base.Joiner;
 import com.google.common.collect.Lists;
 import com.google.common.io.Files;
 import com.google.gson.Gson;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 
+import cromwell.Main;
 import io.cwl.avro.CWL;
 import io.dockstore.client.Bridge;
+import io.dockstore.client.cli.Client;
+import io.dockstore.common.WDLFileProvisioning;
+import io.github.collaboratory.LauncherCWL;
 import io.swagger.client.model.Label;
 import io.swagger.client.ApiException;
 import io.swagger.client.model.SourceFile;
@@ -47,6 +63,7 @@ import static io.dockstore.client.cli.ArgumentUtility.LAUNCH;
 import static io.dockstore.client.cli.ArgumentUtility.CONVERT;
 import static io.dockstore.client.cli.ArgumentUtility.containsHelpRequest;
 import static io.dockstore.client.cli.ArgumentUtility.errorMessage;
+import static io.dockstore.client.cli.ArgumentUtility.exceptionMessage;
 import static io.dockstore.client.cli.ArgumentUtility.invalid;
 import static io.dockstore.client.cli.ArgumentUtility.optVal;
 import static io.dockstore.client.cli.ArgumentUtility.optVals;
@@ -55,6 +72,8 @@ import static io.dockstore.client.cli.ArgumentUtility.printHelpFooter;
 import static io.dockstore.client.cli.ArgumentUtility.printHelpHeader;
 import static io.dockstore.client.cli.ArgumentUtility.reqVal;
 import static io.dockstore.client.cli.Client.CLIENT_ERROR;
+import static io.dockstore.client.cli.Client.IO_ERROR;
+import static io.dockstore.client.cli.Client.API_ERROR;
 
 /**
  * Handles the commands for a particular type of entry. (e.g. Workflows, Tools) Not a great abstraction, but enforces some structure for
@@ -71,6 +90,7 @@ import static io.dockstore.client.cli.Client.CLIENT_ERROR;
  */
 public abstract class AbstractEntryClient {
     protected final CWL cwlUtil = new CWL();
+    private String configFile = null;
 
     /**
      * Print help for this group of commands.
@@ -183,6 +203,9 @@ public abstract class AbstractEntryClient {
                 break;
             case "convert":
                 convert(args);
+                break;
+            case LAUNCH:
+                launch(args);
                 break;
             default:
                 return false;
@@ -484,6 +507,159 @@ public abstract class AbstractEntryClient {
         }
     }
 
+    /**
+     * TODO: this may need to be moved to ToolClient depending on whether we can re-use
+     * this for workflows.
+     * @param args
+     */
+    private void launch(final List<String> args) {
+        if (args.isEmpty() || containsHelpRequest(args)) {
+            launchHelp();
+        } else {
+            final String descriptor = optVal(args, "--descriptor", CWL_STRING);
+
+            String userHome = System.getProperty("user.home");
+            this.configFile = optVal(args, "--config", userHome + File.separator + ".dockstore" + File.separator + "config");
+
+            if (descriptor.equals(CWL_STRING)) {
+                try {
+                    launchCwl(args);
+                } catch (ApiException e) {
+                    exceptionMessage(e, "api error launching workflow", Client.API_ERROR);
+                } catch (IOException e) {
+                    exceptionMessage(e, "io error launching workflow", IO_ERROR);
+                }
+            } else if (descriptor.equals(WDL_STRING)) {
+                launchWdl(args);
+            }
+        }
+    }
+
+    private void launchCwl(final List<String> args) throws ApiException, IOException {
+        final String entry = reqVal(args, "--entry");
+        final String jsonRun = optVal(args, "--json", null);
+        final String csvRuns = optVal(args, "--tsv", null);
+
+        final File tempDir = Files.createTempDir();
+        final SourceFile cwlFromServer = getDescriptorFromServer(entry, "cwl");
+        final File tempCWL = File.createTempFile("temp", ".cwl", tempDir);
+        Files.write(cwlFromServer.getContent(), tempCWL, StandardCharsets.UTF_8);
+
+        downloadDescriptors(entry, "cwl", tempDir);
+
+        final Gson gson = io.cwl.avro.CWL.getTypeSafeCWLToolDocument();
+        if (jsonRun != null) {
+            // if the root document is an array, this indicates multiple runs
+            JsonParser parser = new JsonParser();
+            final JsonElement parsed = parser.parse(new InputStreamReader(new FileInputStream(jsonRun), StandardCharsets.UTF_8));
+            if (parsed.isJsonArray()) {
+                final JsonArray asJsonArray = parsed.getAsJsonArray();
+                for (JsonElement element : asJsonArray) {
+                    final String finalString = gson.toJson(element);
+                    final File tempJson = File.createTempFile("temp", ".json", Files.createTempDir());
+                    FileUtils.write(tempJson, finalString);
+                    final LauncherCWL cwlLauncher = new LauncherCWL(configFile, tempCWL.getAbsolutePath(), tempJson.getAbsolutePath(),
+                            System.out, System.err);
+                    cwlLauncher.run();
+                }
+            } else {
+                final LauncherCWL cwlLauncher = new LauncherCWL(configFile, tempCWL.getAbsolutePath(), jsonRun, System.out, System.err);
+                cwlLauncher.run();
+            }
+        } else if (csvRuns != null) {
+            final File csvData = new File(csvRuns);
+            try (CSVParser parser = CSVParser.parse(csvData, StandardCharsets.UTF_8, CSVFormat.DEFAULT.withDelimiter('\t').withEscape('\\')
+                    .withQuoteMode(QuoteMode.NONE))) {
+                // grab header
+                final Iterator<CSVRecord> iterator = parser.iterator();
+                final CSVRecord headers = iterator.next();
+                // ignore row with type information
+                iterator.next();
+                // process rows
+                while (iterator.hasNext()) {
+                    final CSVRecord csvRecord = iterator.next();
+                    final File tempJson = File.createTempFile("temp", ".json", Files.createTempDir());
+                    StringBuilder buffer = new StringBuilder();
+                    buffer.append("{");
+                    for (int i = 0; i < csvRecord.size(); i++) {
+                        buffer.append("\"").append(headers.get(i)).append("\"");
+                        buffer.append(":");
+                        // if the type is an array, just pass it through
+                        buffer.append(csvRecord.get(i));
+
+                        if (i < csvRecord.size() - 1) {
+                            buffer.append(",");
+                        }
+                    }
+                    buffer.append("}");
+                    // prettify it
+                    JsonParser prettyParser = new JsonParser();
+                    JsonObject json = prettyParser.parse(buffer.toString()).getAsJsonObject();
+                    final String finalString = gson.toJson(json);
+
+                    // write it out
+                    FileUtils.write(tempJson, finalString);
+
+                    // final String stringMapAsString = gson.toJson(stringMap);
+                    // Files.write(stringMapAsString, tempJson, StandardCharsets.UTF_8);
+                    final LauncherCWL cwlLauncher = new LauncherCWL(configFile, tempCWL.getAbsolutePath(), tempJson.getAbsolutePath(),
+                            System.out, System.err);
+                    cwlLauncher.run();
+                }
+            }
+        } else {
+            errorMessage("Missing required parameters, one of  --json or --tsv is required", CLIENT_ERROR);
+        }
+
+    }
+
+    private void launchWdl(final List<String> args) {
+        final String entry = reqVal(args, "--entry");
+        final String json = reqVal(args, "--json");
+
+        Main main = new Main();
+        File parameterFile = new File(json);
+
+        final SourceFile wdlFromServer;
+        try {
+            // Grab WDL from server and store to file
+            wdlFromServer = getDescriptorFromServer(entry, "wdl");
+            final File tempDir = Files.createTempDir();
+            final File tempWdl = File.createTempFile("temp", ".wdl", tempDir);
+            Files.write(wdlFromServer.getContent(), tempWdl, StandardCharsets.UTF_8);
+            downloadDescriptors(entry, "wdl", tempDir);
+
+            // Get list of input files
+            Bridge bridge = new Bridge();
+            Map<String, String> wdlInputs = bridge.getInputFiles(tempWdl);
+
+            // Convert parameter JSON to a map
+            WDLFileProvisioning wdlFileProvisioning = new WDLFileProvisioning(configFile);
+            Gson gson = new Gson();
+            String jsonString = FileUtils.readFileToString(parameterFile);
+            Map<String, Object> map = new HashMap<>();
+            Map<String, Object> inputJson = gson.fromJson(jsonString, map.getClass());
+
+            // Download files and change to local location
+            // Make a new map of the inputs with updated locations
+            Map<String, Object> fileMap = wdlFileProvisioning.pullFiles(inputJson, wdlInputs);
+
+            // Make new json file
+            String newJsonPath = wdlFileProvisioning.createUpdatedInputsJson(inputJson, fileMap);
+
+            final List<String> wdlRun = Lists.newArrayList(newJsonPath, parameterFile.getAbsolutePath());
+            final scala.collection.immutable.List<String> wdlRunList = scala.collection.JavaConversions.asScalaBuffer(wdlRun).toList();
+
+            // run a workflow
+            final int run = main.run(wdlRunList);
+
+        } catch (ApiException ex) {
+            exceptionMessage(ex, "", API_ERROR);
+        } catch (IOException ex) {
+            exceptionMessage(ex, "", IO_ERROR);
+        }
+    }
+
     protected abstract void downloadDescriptors(String entry, String descriptor, File tempDir);
 
     protected String runString(List<String> args, final boolean json) throws
@@ -714,6 +890,24 @@ public abstract class AbstractEntryClient {
         out("Required parameters:");
         out("  --entry <entry>                Complete " + getEntryType().toLowerCase() + " path in the Dockstore");
         out("  --descriptor <descriptor>      Type of descriptor language used. Defaults to cwl");
+        printHelpFooter();
+    }
+
+    private void launchHelp() {
+        printHelpHeader();
+        out("Usage: dockstore " + getEntryType().toLowerCase() + " launch --help");
+        out("       dockstore " + getEntryType().toLowerCase() + " launch [parameters]");
+        out("");
+        out("Description:");
+        out("  Launch an entry locally.");
+        out("");
+        out("Required parameters:");
+        out("  --entry <entry>                     Complete entry path in the Dockstore");
+        out("");
+        out("Optional parameters:");
+        out("  --json <json file>                  Parameters to the entry in the dockstore, one map for one run, an array of maps for multiple runs");
+        out("  --tsv <tsv file>                    One row corresponds to parameters for one run in the dockstore (Only for CWL)");
+        out("  --descriptor <descriptor type>      Descriptor type used to launch workflow. Defaults to " + CWL_STRING);
         printHelpFooter();
     }
